Move a method
